Experiment with storing current server data in redis hash, as a kind of dedupe to help prevent repeated lookups.
Right now (ng builds) this happens: server -> telemetry -> accumulate -> redis -> aggregate -> mongo.

The accumulate step involves a lot of redis hits for no major benefit.

This is what I'm going to be trying instead: server -> telemetry -> redis -> accumulate -> aggregate -> mongo.

The difference is subtle, but it means it should avoid a huge number of redis hits when doing this and other changes, like storing plugin versions used on a server in a redis set instead of computing changes on the fly.

The accumulate step hasn't been written just yet.