Optimize getMissingColumns. On test data, it reduced generation time from 20 secs to 4 secs.
I figured this would have caused issues at some point but the first implementation was really simple to do :-) Either list/set might contain _lots_ of columns so this turns out to be quite expensive when done on every graph!

TreeSet is used as the following 2 conditions need to hold:

1. Column names need to stay the same case when returned
2. Column names in knownColumns and columnsToFetch might have different casing.

This was the simplest way I could think to do it.